pyproject.toml
[project]
name = "site"
version = "0.0.1"
description = "My personal site"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "jx>=0.5.1",
    "markdown>=3.9",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.40.0",
]

[tool.uv.workspace]
members = [
    "apps/blog",
    "apps/portfolio",
    "apps/admin",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
    "taskipy>=1.14.1",
    "pytest-cov>=7.0.0",
    "honcho>=2.0.0",
    "rumdl>=0.1.14",
]

[tool.taskipy.tasks]
fmt = { cmd = "uv run ruff format .", help = "Format code using Ruff." }
lint = { cmd = "uv run ruff check .", help = "Run Ruff lint checks." }
lint-fix = { cmd = "uv run ruff check . --fix", help = "Run Ruff lint checks with autofix." }
test = { cmd = "uv run pytest -q --cov=apps --cov-report=term-missing", help = "Run tests with coverage." }
dev = {cmd = "uv run honcho -f Procfile.dev start", help = "Run dev services with honcho"}

[tool.pytest.ini_options]
testpaths = ["tests"]
